backend/requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    ./bin/compile_dependencies.sh
#
asgiref==3.7.2
    # via django
async-timeout==4.0.2
    # via redis
attrs==23.1.0
    # via
    #   glom
    #   jsonschema
    #   referencing
boltons==23.0.0
    # via
    #   face
    #   glom
certifi==2023.7.22
    # via
    #   django-simple-certmanager
    #   elastic-apm
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via cryptography
charset-normalizer==3.2.0
    # via requests
cryptography==41.0.3
    # via
    #   django-simple-certmanager
    #   josepy
    #   mozilla-django-oidc
    #   pyopenssl
django==3.2.20
    # via
    #   -r requirements/base.in
    #   django-admin-index
    #   django-axes
    #   django-cors-headers
    #   django-formtools
    #   django-hijack
    #   django-otp
    #   django-phonenumber-field
    #   django-privates
    #   django-redis
    #   django-relativedelta
    #   django-rosetta
    #   django-sendfile2
    #   django-simple-certmanager
    #   django-solo
    #   djangorestframework
    #   drf-spectacular
    #   maykin-django-two-factor-auth
    #   mozilla-django-oidc
    #   mozilla-django-oidc-db
    #   zgw-consumers
django-admin-index==3.1.0
    # via -r requirements/base.in
django-axes==6.1.0
    # via -r requirements/base.in
django-better-admin-arrayfield==1.4.2
    # via
    #   -r requirements/base.in
    #   mozilla-django-oidc-db
django-cors-headers==4.2.0
    # via -r requirements/base.in
django-formtools==2.4.1
    # via maykin-django-two-factor-auth
django-hijack==3.4.1
    # via -r requirements/base.in
django-ordered-model==3.7.4
    # via django-admin-index
django-otp==1.2.2
    # via maykin-django-two-factor-auth
django-phonenumber-field==5.2.0
    # via maykin-django-two-factor-auth
django-privates==2.0.0.post0
    # via django-simple-certmanager
django-redis==5.3.0
    # via -r requirements/base.in
django-relativedelta==2.0.0
    # via zgw-consumers
django-rosetta==0.9.9
    # via -r requirements/base.in
django-sendfile2==0.7.0
    # via django-privates
django-simple-certmanager==1.3.0
    # via zgw-consumers
django-solo==2.1.0
    # via
    #   -r requirements/base.in
    #   mozilla-django-oidc-db
    #   zgw-consumers
djangorestframework==3.14.0
    # via
    #   -r requirements/base.in
    #   drf-spectacular
djangorestframework-camel-case==1.4.2
    # via -r requirements/base.in
drf-spectacular==0.26.4
    # via -r requirements/base.in
ecs-logging==2.0.2
    # via elastic-apm
elastic-apm==6.18.0
    # via -r requirements/base.in
face==20.1.1
    # via glom
faker==19.3.0
    # via zgw-consumers
gemma-zds-client==2.0.0
    # via zgw-consumers
glom==23.3.0
    # via mozilla-django-oidc-db
idna==3.4
    # via requests
inflection==0.5.1
    # via drf-spectacular
josepy==1.13.0
    # via mozilla-django-oidc
jsonschema==4.19.0
    # via drf-spectacular
jsonschema-specifications==2023.7.1
    # via jsonschema
maykin-django-two-factor-auth==2.0.4
    # via -r requirements/base.in
mozilla-django-oidc==2.0.0
    # via mozilla-django-oidc-db
mozilla-django-oidc-db==0.12.0
    # via -r requirements/base.in
phonenumbers==8.13.18
    # via -r requirements/base.in
pillow==10.0.0
    # via -r requirements/base.in
polib==1.2.0
    # via django-rosetta
psycopg2==2.9.7
    # via -r requirements/base.in
pycparser==2.21
    # via cffi
pyjwt==2.8.0
    # via gemma-zds-client
pyopenssl==23.2.0
    # via
    #   django-simple-certmanager
    #   josepy
    #   zgw-consumers
python-dateutil==2.8.2
    # via
    #   django-relativedelta
    #   faker
python-decouple==3.8
    # via -r requirements/base.in
python-dotenv==1.0.0
    # via -r requirements/base.in
pytz==2023.3
    # via
    #   -r requirements/base.in
    #   django
    #   djangorestframework
pyyaml==6.0.1
    # via
    #   drf-spectacular
    #   gemma-zds-client
qrcode==6.1
    # via maykin-django-two-factor-auth
redis==4.6.0
    # via django-redis
referencing==0.30.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   django-rosetta
    #   gemma-zds-client
    #   mozilla-django-oidc
    #   requests-mock
    #   zgw-consumers
requests-mock==1.11.0
    # via zgw-consumers
rpds-py==0.9.2
    # via
    #   jsonschema
    #   referencing
sentry-sdk==1.29.2
    # via -r requirements/base.in
six==1.16.0
    # via
    #   python-dateutil
    #   qrcode
    #   requests-mock
sqlparse==0.4.4
    # via django
typing-extensions==4.7.1
    # via asgiref
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.0.4
    # via
    #   elastic-apm
    #   requests
    #   sentry-sdk
uwsgi==2.0.22
    # via -r requirements/base.in
wrapt==1.15.0
    # via elastic-apm
zgw-consumers==0.26.2
    # via -r requirements/base.in

# The following packages are considered to be unsafe in a requirements file:
setuptools==68.0.0
    # via
    #   django-axes
    #   josepy
